fix(wamn-fz4d): the WMS and Acme revisions are int32, so routes and codecs agree
The Acme quality approve route refused every request. Its route schema
said integer, but its codec read the int64 revision only from a JSON
string. The WMS pallet revision was the other int64 revision left
(wamn-kaxj). Under the integer ruling, both become int32.

Source:
- The WMS pallet row_version and the four command claim columns that
  record it are int4. The Acme quality_inspection row_version is int4.
  Both migrations are edited in place under the POC rule.
- Every int64 in both manifests was a revision, and each one is int32:
  30 in WMS and 5 in Acme.
- The WMS and Acme data and component Rust narrows from i64 to i32.
- The WMS component declaration, the four inventory route schemas, and
  the move result of the label wiring take an integer revision. The four
  attachment definition hashes move. The Acme route schemas already
  said integer, and now its codec agrees.

Regenerated WMS and Acme. SQLx metadata was prepared again for both.

Callers and tests send and read numbers. The WMS move example, the
PTY and preflight scripts, the negative control and the WMS live
contention test are updated. local_business seeds the revision
2_147_483_547, near the top of int4, where it seeded 2^53 + 1 to test
exact int64 carriage. The Receiving journey tests are updated too.
They sent the string "1" for the int32 purchase order revision, which
the route schema refuses, and two of them run in the Acme cluster case.
The demo page holds the pallet revision as a number.

Tests run: Acme publication and SQLx tests; WMS publication and SQLx
tests; local_business::operations_and_replay and
route_interface_live::a_route_answers_every_operation_kind_as_its_one_node_wiring
over freshly built guests; Receiving library tests (32 passed, 19
ignored cluster cases not run); the WMS move example; clippy and fmt.
The Acme cluster case is not run here.

Closes wamn-kaxj.

+++ apps/client_acme_receiving/component/src/lib.rs
@@ -137,7 +137,7 @@ fn access_detail(
     key: &str,
     operation: &str,
     not_found: Option<(&str, &str)>,
-    expected_row_version: Option<i64>,
+    expected_row_version: Option<i32>,
 ) -> Option<String> {
     match key {
         "field" => error.field().map(str::to_owned).or_else(|| {
