+++ apps/client_acme_receiving/data/src/error.rs
@@ -53,7 +53,7 @@ pub struct AccessError {
     context: Box<str>,
     constraint: Option<Box<str>>,
     field: Option<&'static str>,
-    observed_row_version: Option<i64>,
+    observed_row_version: Option<i32>,
 }
 
 impl AccessError {
@@ -78,7 +78,7 @@ impl AccessError {
     }
 
     /// Current revision returned by an optimistic-concurrency refusal.
-    pub const fn observed_row_version(&self) -> Option<i64> {
+    pub const fn observed_row_version(&self) -> Option<i32> {
         self.observed_row_version
     }
 
@@ -98,7 +98,7 @@ impl AccessError {
 
     pub(crate) fn concurrency_conflict(
         context: impl Into<Box<str>>,
-        observed_row_version: i64,
+        observed_row_version: i32,
     ) -> Self {
         Self {
             kind: AccessErrorKind::ConcurrencyConflict,
